Time: 0 ms (100%), Space: 17.7 MB (13.49%) - LeetHub

+++ 0801-is-graph-bipartite/0801-is-graph-bipartite.cpp
@@ -62,7 +62,7 @@ public:
     }
 
     bool _isBipartite(int n, const vector<vector<int>> &adjList){
-        vector<int> colors(n, -1);
+        vector<int8_t> colors(n, -1);
         for(int node = 1; node < n; node++){
             if(colors[node] != -1){
                 continue;
